# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API",
        "description": "Documentação apresentando todos endpoints desta API. Contém informações de autenticação, corpo da requisição e respostas.",
        "contact": {
            "email": "hiagolaureano@hotmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/v1/users": {
            "get": {
                "tags": [
                    "/v1/users"
                ],
                "summary": "Obter lista de usuários",
                "description": "Retorna uma lista com todos usuários",
                "operationId": "1e366e9046d9b43b4c482a9f5a8f3f46",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da paginação dos dados. São 15 registros por paginação",
                        "style": "form",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Dados da requisição",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "links": {
                                            "description": "Links da paginação",
                                            "type": "object",
                                            "example": {
                                                "first": "http:localhost/api/v1/item?page=1",
                                                "last": "http:localhost/api/v1/item?page=3",
                                                "prev": "http:localhost/api/v1/item?page=1",
                                                "next": "http:localhost/api/v1/item?page=3"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Login necessário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Somente superusuários possuem essa permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "/v1/users"
                ],
                "summary": "Registrar novo usuário",
                "description": "Cria um novo usuário",
                "operationId": "9e68102db783652fa512181fa8cb21b7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "phone",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "first_name": {
                                        "description": "Nome",
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "description": "Sobrenome",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Número de telefone (Somente números, código do país + DDD + número)",
                                        "type": "string",
                                        "pattern": "5544999998888"
                                    },
                                    "email": {
                                        "description": "E-mail",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Senha",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Novo registro criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Login necessário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Somente o propretário do registro e superusuários possuem essa permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "420": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "pattern": "O campo x é obrigatório. (+ 3 erros)"
                                        },
                                        "errors": {
                                            "description": "Campos com erro de validação e lista dos erros",
                                            "type": "object",
                                            "example": {
                                                "campo_x": [
                                                    "O campo x é obrigatório."
                                                ],
                                                "campo_y": [
                                                    "O campo y deve ser um inteiro."
                                                ],
                                                "campo_z": [
                                                    "O campo z deve ter no máximo 30 caracteres."
                                                ],
                                                "campo_w": [
                                                    "O campo w é obrigatório."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/users/{id}": {
            "get": {
                "tags": [
                    "/v1/users"
                ],
                "summary": "Obter usuário específico",
                "description": "Retorna um usuário específico",
                "operationId": "7f33edb5bcfdf9c57a2058abca7760e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id do usuário",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Login necessário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Somente o propretário do registro e superusuários possuem essa permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "/v1/users"
                ],
                "summary": "Deletar usuário",
                "description": "Deleta um usuário",
                "operationId": "baf7273016f7db2a04dfad84dc8b1028",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id do usuário",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sucesso, sem corpo de resposta"
                    },
                    "401": {
                        "description": "Login necessário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Somente o propretário do registro e superusuários possuem essa permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User Model",
                "properties": {
                    "id": {
                        "description": "Id do usuário",
                        "type": "integer"
                    },
                    "first_time": {
                        "description": "Nome",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "Sobrenome",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Número de telefone",
                        "type": "integer",
                        "pattern": "5500999990000"
                    },
                    "email": {
                        "description": "E-mail",
                        "type": "string"
                    },
                    "email_verified_at": {
                        "description": "Data de verificação do e-mail",
                        "type": "string",
                        "pattern": "20/12/2024 11:12:40",
                        "nullable": true
                    },
                    "is_staff": {
                        "description": "Se é membro da equipe",
                        "type": "boolean"
                    },
                    "is_superuser": {
                        "description": "Se é superusuário",
                        "type": "boolean"
                    },
                    "pets": {
                        "description": "Lista com todos os dados dos pets do usuário",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Pet"
                        }
                    },
                    "created_at": {
                        "description": "Data da criação do registro",
                        "type": "string",
                        "pattern": "20/12/2024 11:12:40"
                    },
                    "updated_at": {
                        "description": "Data da atualização do registro",
                        "type": "string",
                        "pattern": "21/12/2024 11:15:30"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}